+++ packages/cli/src/__fixtures__/extension-removal/expect-sync/src/locale/ja.ts
@@ -1,7 +1,7 @@
 import { Catalog, msg } from "@hi18n/core";
 import type { Vocabulary } from ".";
 
-export default new Catalog<Vocabulary>({
+export default new Catalog<Vocabulary>("ja", {
   "example/dynamic-todo": msg.todo("[TODO: example/dynamic-todo]"),
   "example/greeting": msg("こんにちは世界!"),
   "example/greeting-todo-newtranslation": msg.todo("[TODO: example/greeting-todo-newtranslation]"),